backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Async Support
aiofiles==23.2.1
httpx==0.25.2
aiohttp==3.9.1

# PDF Processing
beautifulsoup4[lxml]==4.12.2
lxml==4.9.3

# Vector Search
faiss-cpu==1.7.4
numpy==1.26.2

# OpenAI Integration
openai==1.3.5
tiktoken==0.5.1

# Configuration and Type Hints
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
typing-extensions>=4.8.0

# HTTP Client
requests==2.31.0

# XML Processing
xmltodict==0.13.0 